2.1.16

+++ atec-cache-apcu/trunk/includes/atec-cache-apcu-settings.php
@@ -8,8 +8,8 @@ $wpc_tools=new ATEC_wpc_tools();
 require_once('atec-check.php');
 
 global $atec_wpca_apcu_enabled;
-$optionName='atec_WPCA_settings';
-$options=get_option($optionName,[]);
+$optName='atec_WPCA_settings';
+$options=get_option($optName,[]);
 $atec_wpca_pcache = $options['cache']??null==true;
 
 $advanced=defined('WP_APCU_MU_PAGE_CACHE');
@@ -50,25 +50,19 @@ echo '
 		<div class="atec-border-white">
 		<h4>'.esc_attr__('APCu Page Cache','atec-cache-apcu').' '; $wpc_tools->enabled($atec_wpca_pcache); echo '</h4><hr>';
 							
-		if ($atec_wpca_apcu_enabled)
+		if ($atec_wpca_apcu_enabled && class_exists('APCUIterator'))
 		{    
-			$apcu_cache=apcu_cache_info(true);
-			if ($apcu_cache)
+			if (!empty($apcu_it=new APCUIterator('/atec_WPCA_p_/')))
 			{
-				$apcu_it=new APCUIterator('/atec_WPCA_p_/');
-				if (!empty($apcu_it))
-				{
-					$c=0; $size=0;
-					foreach ($apcu_it as $entry) 
-					{ if (!str_contains($entry['key'],'_h')) { $c++; $size+=$entry['mem_size']; }
-					}
-					echo '<p style="padding: 4px;" class="atec-box-white">', 
-					esc_attr__('Current size is','atec-cache-apcu'), ' <strong>', esc_attr(size_format($size)),
-					'</strong> (', esc_attr(number_format($c)), ' ', $c>1?esc_attr__('items','atec-cache-apcu'):esc_attr__('item','atec-cache-apcu'), 	').</p>';
-				}
-				else { $wpc_tools->error('',__('No page cache data available','atec-cache-apcu')); }
-				echo '<p>'.esc_attr__('The page cache is an additional feature of this plugin','atec-cache-apcu').'.<br>'.esc_attr__('It will give your page an additonal boost, by delivering pages from APCu cache','atec-cache-apcu').'.</p>';
+				$c=0; $size=0;
+				foreach ($apcu_it as $entry) 
+				{ if (!str_contains($entry['key'],'_h')) { $c++; $size+=$entry['mem_size']; } }
+				echo '<p style="padding: 4px;" class="atec-box-white">', 
+				esc_attr__('Current size is','atec-cache-apcu'), ' <strong>', esc_attr(size_format($size)),
+				'</strong> (', esc_attr(number_format($c)), ' ', $c>1?esc_attr__('items','atec-cache-apcu'):esc_attr__('item','atec-cache-apcu'), 	').</p>';
 			}
+			else { $wpc_tools->error('',__('No page cache data available','atec-cache-apcu')); }
+			echo '<p>'.esc_attr__('The page cache is an additional feature of this plugin','atec-cache-apcu').'.<br>'.esc_attr__('It will give your page an additonal boost, by delivering pages from APCu cache','atec-cache-apcu').'.</p>';
 		}
 		else { $wpc_tools->error('',__('APCu not enabled – Page cache disabled','atec-cache-apcu').'.'); }
 		echo '<p>', esc_attr__('The page cache saves pages, posts and categories – no product/shop pages (WooCommerce).','atec-cache-apcu'), '</p>';
